Find Characters
Given a list and a character within that list print only the words that have the given character in it.  For example I have a list with [apple, dog, rabbit] and a character ["o"]. If the function is written correctly it will print out only dog in a new list.