Fun with Functions
It's a two-in-one type of function. The first, Odd+Even, is to print each number with its corresponding partner of even or odd. The second is to multiply a given list with five and then print out that new list. 

I did not do the hacker challenge but plan and coming back to it. 

EDIT: Also, the one thing that I couldn't figure out but maybe a VERY simple fix I am missing is that in the range I am always one digit short. For example: I put in a range of 1,50 but it will only print out to 1-49.