Scores and Grades
This exercise involves creating a function that would grade 10 random numbers given. So, it would essentially list out 10 scores and grades randomly. 
This one in particular gives the benefit of the doubt to the poor students! Or pessimistic depending on your view of glasses. The grades will never fall beneath 60. 
I will create a new modified one that will grade on a lower score.