Coin Toss
This one was, on paper, deceptively easy but tricks you a little bit. But I got it figured out and had it finished in no time. Also, as a note, if anyone really reads these commits. The reason I added "+1" to the x is so that it would accurately count the actual number of tosses.