Making Dictionary
"Create a function that takes in two lists and creates a single dictionary. The first list contains keys and the second list contains the values. Assume the lists will be of equal length."

The "hacker challenge" would be to change the position of key and values depending on the length of the values.